+++ packages/google/src/google-generative-ai-options.ts
@@ -30,9 +30,6 @@ export type GoogleGenerativeAIModelId =
   | 'gemini-2.5-flash'
   | 'gemini-2.5-flash-image-preview'
   | 'gemini-2.5-flash-lite'
-  | 'gemini-2.5-flash-lite-preview-09-2025'
-  | 'gemini-2.5-flash-preview-04-17'
-  | 'gemini-2.5-flash-preview-09-2025'
   | 'gemini-2.5-flash-preview-tts'
   | 'gemini-2.5-pro-preview-tts'
   | 'gemini-2.5-flash-native-audio-latest'
@@ -45,6 +42,8 @@ export type GoogleGenerativeAIModelId =
   | 'gemini-3.1-pro-preview'
   | 'gemini-3.1-pro-preview-customtools'
   | 'gemini-3.1-flash-image-preview'
+  | 'gemini-3.1-flash-lite-preview'
+  | 'gemini-3.1-flash-tts-preview'
   // latest version
   // https://ai.google.dev/gemini-api/docs/models#latest
   | 'gemini-pro-latest'