+++ packages/ai/CHANGELOG.md
@@ -1,5 +1,46 @@
 # ai
 
+## 5.0.179
+
+### Patch Changes
+
+- 5543cd1: Add AI Gateway hint to provider READMEs
+
+## 5.0.178
+
+### Patch Changes
+
+- Updated dependencies [eacfb5e]
+  - @ai-sdk/gateway@2.0.82
+
+## 5.0.177
+
+### Patch Changes
+
+- Updated dependencies [dc91bcc]
+  - @ai-sdk/gateway@2.0.81
+
+## 5.0.176
+
+### Patch Changes
+
+- Updated dependencies [4b40033]
+  - @ai-sdk/gateway@2.0.80
+
+## 5.0.175
+
+### Patch Changes
+
+- Updated dependencies [d4fbea5]
+  - @ai-sdk/gateway@2.0.79
+
+## 5.0.174
+
+### Patch Changes
+
+- Updated dependencies [140640d]
+  - @ai-sdk/gateway@2.0.78
+
 ## 5.0.173
 
 ### Patch Changes